rename: populationCompositionをpopulationCompに省略すると決めたのでそれに伴うファイル名や関数名の変更
