+++ src/__tests__/services/api/populationComp/fetchPopulationComp.test.ts
@@ -1,7 +1,7 @@
-import fetchPopulationComposition from '@/services/api/populationComposition/fetchPopulationComposition';
-import { PopulationCompositionResponse } from '@/types/api/models/populationComposition/PopulationCompositionResponse';
+import fetchPopulationComp from '@/services/api/populationComp/fetchPopulationComp';
+import { PopulationCompResponse } from '@/types/api/models/populationComp/PopulationCompoResponse';
 
-describe('fetchPopulationComposition', () => {
+describe('fetchPopulationComp', () => {
   // 元のfetchを保存する変数
   let originalFetch: typeof global.fetch;
 
@@ -26,7 +26,7 @@ describe('fetchPopulationComposition', () => {
     // モックのfetch関数を定義
     global.fetch = jest.fn().mockResolvedValue({
       ok: true,
-      json: async (): Promise<{ result: PopulationCompositionResponse }> => ({
+      json: async (): Promise<{ result: PopulationCompResponse }> => ({
         result: {
           boundaryYear: 2020,
           data: [
@@ -48,12 +48,12 @@ describe('fetchPopulationComposition', () => {
       }),
     });
 
-    // fetchPopulationComposition関数を実行
-    const response: PopulationCompositionResponse | false = await fetchPopulationComposition({
+    // fetchPopulationComp関数を実行
+    const response: PopulationCompResponse | false = await fetchPopulationComp({
       prefCode: 1,
     });
 
-    // fetchPopulationComposition関数の期待する戻り値と一致するか確認
+    // fetchPopulationComp関数の期待する戻り値と一致するか確認
     expect(response).toEqual({
       boundaryYear: 2020,
       data: [
@@ -101,7 +101,7 @@ describe('fetchPopulationComposition', () => {
     });
     global.fetch = mockFetch;
 
-    await fetchPopulationComposition({ prefCode: 1 });
+    await fetchPopulationComp({ prefCode: 1 });
 
     // 正しいURLでfetchが呼ばれているか
     expect(mockFetch).toHaveBeenCalledWith(
@@ -128,8 +128,8 @@ describe('fetchPopulationComposition', () => {
       statusText: 'Internal Server Error',
     });
 
-    // fetchPopulationComposition関数を実行
-    const response: PopulationCompositionResponse | false = await fetchPopulationComposition({
+    // fetchPopulationComp関数を実行
+    const response: PopulationCompResponse | false = await fetchPopulationComp({
       prefCode: 1,
     });
 
@@ -145,8 +145,8 @@ describe('fetchPopulationComposition', () => {
       json: async (): Promise<unknown> => ({ message: 'No result' }),
     });
 
-    // fetchPopulationComposition関数を実行
-    const response: PopulationCompositionResponse | false = await fetchPopulationComposition({
+    // fetchPopulationComp関数を実行
+    const response: PopulationCompResponse | false = await fetchPopulationComp({
       prefCode: 1,
     });
 
@@ -159,8 +159,8 @@ describe('fetchPopulationComposition', () => {
     // モックのfetch関数を定義
     global.fetch = jest.fn().mockRejectedValue(new Error('Network error'));
 
-    // fetchPopulationComposition関数を実行
-    const response: PopulationCompositionResponse | false = await fetchPopulationComposition({
+    // fetchPopulationComp関数を実行
+    const response: PopulationCompResponse | false = await fetchPopulationComp({
       prefCode: 1,
     });
