+++ src/conf/api/apiPath.ts
@@ -4,5 +4,5 @@ export const apiPath: ApiPath = {
   // APIから都道府県一覧を取得するパス
   PREFECTURES: '/api/v1/prefectures',
   // APIから人口構成を取得するパス
-  POPULATION_COMPOSITION: '/api/v1/population/composition/perYear',
+  POPULATION_COMP: '/api/v1/population/composition/perYear',
 };
