+++ src/types/api/conf/ApiPath.ts
@@ -2,5 +2,5 @@ export interface ApiPath {
   // APIから都道府県一覧を取得するパス
   PREFECTURES: string;
   // APIから人口構成を取得するパス
-  POPULATION_COMPOSITION: string;
+  POPULATION_COMP: string;
 }
