+++ src/types/api/models/populationComp/PopulationComp.ts
@@ -1,10 +1,10 @@
-import { YearlyPopulationData } from '@/types/api/models/populationComposition/YearlyPopulationData';
+import { YearlyPopulationData } from '@/types/api/models/populationComp/YearlyPopulationData';
 
 /**
  * 人口構成のカテゴリごとのデータ構造
  * 総人口、年少人口、生産年齢人口、老年人口
  */
-export interface PopulationComposition {
+export interface PopulationComp {
   label: string;
   data: YearlyPopulationData[];
 }
