--- src/types/api/models/populationComposition/PopulationCompositionResponse.ts
@@ -1,11 +1,0 @@
-import { PopulationComposition } from '@/types/api/models/populationComposition/PopulationComposition';
-
-/**
- * getPopulationCompositionのレスポンスのデータ構造
- */
-export interface PopulationCompositionResponse {
-  // 実績値と推計値の区切り年
-  boundaryYear: number;
-  // 人口構成のカテゴリごとのデータ
-  data: PopulationComposition[];
-}
